requirements.txt
fastapi
uvicorn
kafka-python
python-dotenv
pydantic
pytest
requests